Rename switcher module to build_status
